warning fixes, one might also want to change hardcoded colors into system colors, but that's later on...
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@24372 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/cortex/ValControl/ValControlDigitSegment.cpp
@@ -316,7 +316,7 @@ void ValControlDigitSegment::MessageReceived(BMessage* pMsg) {
 		case ValControl::M_SET_VALUE:
 			err = pMsg->FindDouble("value", &fVal);
 			ASSERT(err == B_OK);
-			setValue(fVal, fVal < 0);
+			setValue((int64)fVal, fVal < 0);
 			break;
 			
 		case ValControl::M_GET_VALUE: {
